requirements.txt
biopython>=1.81
goatools>=1.3.1
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
click>=8.0.0
pyyaml>=6.0
statsmodels>=0.13.0
seaborn>=0.11.0